requirements.txt
bleach==6.2.0
blinker==1.9.0
click==8.2.1
colorama==0.4.6
dominate==2.9.1
Flask==3.1.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pyodbc==5.2.0
python-dotenv==1.1.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
visitor==0.1.3
webencodings==0.5.1
Werkzeug==3.1.3
WTForms==3.2.1
email_validator
